+++ colossalai/zero/gemini/__init__.py
@@ -1,9 +1,4 @@
 from .chunk import ChunkManager, TensorInfo, TensorState, search_chunk_configuration
 from .gemini_mgr import GeminiManager
-from .stateful_tensor_mgr import StatefulTensorMgr
-from .tensor_placement_policy import TensorPlacementPolicyFactory
 
-__all__ = [
-    'StatefulTensorMgr', 'TensorPlacementPolicyFactory', 'GeminiManager', 'TensorInfo', 'TensorState', 'ChunkManager',
-    'search_chunk_configuration'
-]
+__all__ = ['GeminiManager', 'TensorInfo', 'TensorState', 'ChunkManager', 'search_chunk_configuration']
