+++ colossalai/zero/legacy/gemini/__init__.py
@@ -1,0 +1,13 @@
+from .stateful_tensor import StatefulTensor
+from .stateful_tensor_mgr import StatefulTensorMgr
+from .tensor_placement_policy import AutoTensorPlacementPolicy, CPUTensorPlacementPolicy, CUDATensorPlacementPolicy
+from .zero_hook import ZeroHook
+
+__all__ = [
+    'StatefulTensorMgr',
+    'StatefulTensor',
+    'CPUTensorPlacementPolicy',
+    'CUDATensorPlacementPolicy',
+    'AutoTensorPlacementPolicy',
+    'ZeroHook',
+]
